remove old cmds
