--- packages/@tinacms/cli/src/cmds/compile/defaultSchema.ts
@@ -1,81 +1,0 @@
-/**
-
-*/
-
-export const defaultSchema = `
-import { defineSchema, defineConfig } from 'tinacms'
-import { client } from './__generated__/client'
-
-
-const branch =
-  process.env.NEXT_PUBLIC_TINA_BRANCH ||
-  process.env.NEXT_PUBLIC_VERCEL_GIT_COMMIT_REF ||
-  process.env.HEAD ||
-  'main'
-const schema = defineSchema({
-  // See https://tina.io/docs/tina-cloud/connecting-site/ for more information about this config
-  config: {
-    token: '<Your Read Only Token>', // generated on app.tina.io,
-    clientId: '<Your Client ID>', // generated on app.tina.io
-    branch,
-  },
-  collections: [
-    {
-      label: 'Blog Posts',
-      name: 'post',
-      path: 'content/posts',
-      format: 'mdx',
-      ui: {
-        router: ({ document }) => {
-          // This can be used to add contextual editing to your site. See https://tina.io/docs/tinacms-context/#accessing-contextual-editing-from-the-cms for more information.
-          return \`/demo/blog/\${document._sys.filename}\`
-        },
-      },
-      fields: [
-        {
-          type: 'string',
-          label: 'Title',
-          name: 'title',
-        },
-        {
-          type: 'rich-text',
-          label: 'Blog Post Body',
-          name: 'body',
-          isBody: true,
-          templates: [
-            {
-              name: 'PageSection',
-              label: 'Page Section',
-              fields: [
-                {
-                  type: 'string',
-                  name: 'heading',
-                  label: 'Heading',
-                },
-                {
-                  type: 'string',
-                  name: 'content',
-                  label: 'Content',
-                  ui: {
-                    component: 'textarea',
-                  },
-                },
-              ],
-            },
-          ],
-        },
-      ],
-    },
-  ],
-})
-
-export default schema
-
-// Your tina config
-
-export const tinaConfig = defineConfig({
-  client,
-  schema,
-})
-
-`
