--- packages/@tinacms/cli/src/cmds/query-gen/index.ts
@@ -1,5 +1,0 @@
-/**
-
-*/
-
-export * from './genTypes'
