--- packages/@tinacms/cli/src/cmds/start-server/server.ts
@@ -1,6 +1,0 @@
-/**
-
-*/
-
-import { gqlServer } from '../../server'
-export default gqlServer
