--- packages/@tinacms/cli/src/command.ts
@@ -1,20 +1,0 @@
-/**
-
-*/
-
-export interface Command {
-  resource?: string
-  command: string
-  alias?: string
-  description: string
-  action: (...args: any[]) => void
-  examples?: string
-  subCommands?: Command[]
-  options?: Option[]
-}
-
-interface Option {
-  name: string
-  description: string
-  defaultValue?: any
-}
