+++ packages/@tinacms/cli/src/next/commands/dev-command/index.ts
@@ -16,7 +16,7 @@ import { logger, summary } from '../../../logger'
 import { createDBServer, createDevServer } from './server'
 import { Codegen } from '../../codegen'
 import chalk from 'chalk'
-import { startSubprocess2 } from '../../../cmds/startSubprocess'
+import { startSubprocess2 } from '../../../utils/start-subprocess'
 
 export class DevCommand extends Command {
   static paths = [['dev'], ['server:start']]
